Moved CameraCtrl and AnimateLCM-I2V nodes into their own files
